Add CogAt smoke test.
- Reorganize ontology resources.
- Fix up CogAt functions.
- Add test dataset with abstracts (only articles by Angie so no issues
there).
- Add CogAt smoke test.
