+++ nimare/annotate/ontology/__init__.py
@@ -2,7 +2,9 @@
 Automated annotation tools for existing ontologies.
 """
 from .cogpo import extract_cogpo
-from .cogat import extract_cogat, expand_counts, CogAtLemmatizer
+from .cogat import download_cogat, extract_cogat, expand_counts, CogAtLemmatizer
 
-__all__ = ['extract_cogpo', 'extract_cogat', 'expand_counts',
-           'CogAtLemmatizer']
+__all__ = [
+    'extract_cogpo',
+    'download_cogat', 'extract_cogat', 'expand_counts', 'CogAtLemmatizer'
+]
