+++ nimare/resources/ontologies/cognitive_atlas/cogat_aliases.csv
