+++ nimare/resources/ontologies/cognitive_atlas/cogat_relationships.csv
